feat(cli): move from cmd
BREAKING CHANGE: We have broken this a few times and we feel it is time for v2.
