+++ cli/command.go
@@ -1,0 +1,8 @@
+package cli
+
+import "github.com/alexfalkowski/go-service/cli/flag"
+
+// Command for cli.
+type Command struct {
+	*flag.FlagSet
+}
