+++ cli/config/config.go
@@ -4,7 +4,7 @@ import (
 	"errors"
 
 	"github.com/alexfalkowski/go-service/bytes"
-	"github.com/alexfalkowski/go-service/cmd/flag"
+	"github.com/alexfalkowski/go-service/cli/flag"
 	"github.com/alexfalkowski/go-service/config/io"
 	"github.com/alexfalkowski/go-service/encoding"
 	"github.com/alexfalkowski/go-service/env"
