+++ cli/module.go
@@ -1,10 +1,9 @@
-package cmd
+package cli
 
 import "go.uber.org/fx"
 
 // Module for fx.
 var Module = fx.Options(
-	fx.Provide(NewCommand),
 	fx.Provide(NewInputConfig),
 	fx.Provide(NewOutputConfig),
 )
