+++ config/config_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	cache "github.com/alexfalkowski/go-service/cache/config"
-	"github.com/alexfalkowski/go-service/cmd/flag"
+	"github.com/alexfalkowski/go-service/cli/flag"
 	"github.com/alexfalkowski/go-service/config"
 	"github.com/alexfalkowski/go-service/crypto"
 	"github.com/alexfalkowski/go-service/crypto/aes"
